description:
- clean and optimize the code;
- add authentication and authorization checks to routes;
- upload image amazon s3;
- bug fixes;
- docs: add documentation.

+++ api/useCases/DeleteRating/DeleteRatingUseCase.ts
@@ -4,6 +4,10 @@ export class DeleteRatingUseCase {
   constructor(private ratingRepository: IRatingRepository) {}
 
   async execute(id: string) {
-    return this.ratingRepository.deleteRating(id)
+    const ratingExists: boolean = await this.ratingRepository.verifyIdRatingExists(id)
+    if(!ratingExists){
+      throw new Error('Essa avaliação não existe!');
+    }
+    return await this.ratingRepository.deleteRating(id)
   }
 }
